08_semestre/pdytr/practica/tp_3/ejer3/requirements.txt
grpc-stubs==1.53.0.5
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2
setuptools==75.1.0
